Add test one pixel attack

+++ cifar10/config.py
@@ -2,19 +2,19 @@ import numpy as np
 
 test_model_thresholds = {
     "plain": {
-        "l0": np.linspace(0, 15, 16),
+        "l0": np.linspace(0, 15, 16, dtype=np.int32),
         "l1": np.linspace(2, 10, 5),
         "l2": [0.1, 0.15, 0.2, 0.3, 0.4],
         "li": np.linspace(1, 3, 5) / 255.0
     },
     "linf": {
-        "l0": np.linspace(0, 15, 16),
+        "l0": np.linspace(0, 15, 16, dtype=np.int32),
         "l1": np.linspace(5, 20, 5),
         "l2": np.linspace(0.25, 1.25, 5),
         "li": np.linspace(2, 10, 5) / 255.0
     },
     "l2": {
-        "l0": np.linspace(0, 15, 16),
+        "l0": np.linspace(0, 15, 16, dtype=np.int32),
         "l1": np.linspace(3, 15, 5),
         "l2": np.linspace(0.25, 1.25, 5),
         "li": np.linspace(2, 10, 5) / 255.0
