+++ mnist/config.py
@@ -8,13 +8,13 @@ test_model_thresholds = {
         "li": np.linspace(0.03, 0.11, 5)
     },
     "linf": {
-        "l0": np.linspace(2, 10, 5),
+        "l0": np.linspace(2, 10, 5, dtype=np.int32),
         "l1": np.linspace(2.5, 12.5, 5),
         "l2": np.linspace(1.0, 3.0, 5),
         "li": [0.2, 0.25, 0.3, 0.325, 0.35],
     },
     "l2": {
-        "l0": np.linspace(5, 25, 5),
+        "l0": np.linspace(5, 25, 5, dtype=np.int32),
         "l1": np.linspace(5, 20, 5),
         "l2": np.linspace(1.0, 3.0, 5),
         "li": np.linspace(0.05, 0.25, 5),
